Move ReactJsExceptionHandler -> interfaces package (#38241)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38241

In this diff I'm moving the ReactJsExceptionHandler interface into the react.interfaces package.
This bring consistence with the rest of bridgeless APIs

changelog: [internal] internal

Reviewed By: cortinico

Differential Revision: D47277572

fbshipit-source-id: 30e5c7318c93fe04919297790514b4a4c63a9e6a
