+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/interfaces/exceptionmanager/ReactJsExceptionHandler.java
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-package com.facebook.react.bridgeless.exceptionmanager;
+package com.facebook.react.interfaces.exceptionmanager;
 
 import com.facebook.proguard.annotations.DoNotStripAny;
 import com.facebook.react.common.mapbuffer.ReadableMapBuffer;
